+++ src/dreamstate/core/queries/queryData.ts
@@ -5,6 +5,8 @@ import { IOptionalQueryRequest, IQueryRequest, TQueryResponse, TQueryType } from
 /**
  * Find correct async listener or array of listeners and return response or null.
  * Try to find matching type and call related method.
+ *
+ * @deprecated multiple queries at once, only one query per call will remain
  */
 export function queryData<
   R,